% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{BRAVO}
\alias{BRAVO}
\title{BRAVO
This cleans HRM data}
\usage{
BRAVO(x, channel)
}
\arguments{
\item{x}{dataframe}
}
\description{
BRAVO
This cleans HRM data
}
\examples{
BRAVO(x)
}
\keyword{BRAVO}
\keyword{CleanUp}
