% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{dataBRAVO}
\alias{dataBRAVO}
\title{dataBRAVO
Extracts from the BRAVO table
This acquires the first two BRAVO days data frmo the upper GI database}
\usage{
dataBRAVO(channel)
}
\arguments{
\item{channel}{odbcConnectAccess connection defined at the start of this file from RODBC function (for windows)}
}
\description{
dataBRAVO
Extracts from the BRAVO table
This acquires the first two BRAVO days data frmo the upper GI database
}
\examples{
#dataBRAVO(channel)
}
\keyword{BRAVO}
\keyword{Extraction}
