% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{dataBRAVOClean}
\alias{dataBRAVOClean}
\title{dataBRAVOClean
This cleans the BRAVO Data}
\usage{
dataBRAVOClean(x)
}
\arguments{
\item{x}{dataframe usually the standard impedance data}
}
\description{
dataBRAVOClean
This cleans the BRAVO Data
}
\examples{
#dataBRAVOClean(x)
}
\keyword{BRAVO}
\keyword{CleanUp}
