% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{dataBRAVODayLabeller}
\alias{dataBRAVODayLabeller}
\title{dataBRAVODayLabeller
This organises the BRAVO data by day. Currently the files are extracted so that you get two days (usually)
per BRAVO file. This function will associate BRAVOs for one patient separated by max 4 days. The later file will have column names
changed to day3 and day4 and a flat dataframe produced with all the values for one BRAVO 'session' produced
to make downstream analyses easier.}
\usage{
dataBRAVODayLabeller(x, HospNum_Id, VisitDate)
}
\arguments{
\item{x}{dataframe usually the standard impedance data}
}
\description{
dataBRAVODayLabeller
This organises the BRAVO data by day. Currently the files are extracted so that you get two days (usually)
per BRAVO file. This function will associate BRAVOs for one patient separated by max 4 days. The later file will have column names
changed to day3 and day4 and a flat dataframe produced with all the values for one BRAVO 'session' produced
to make downstream analyses easier.
}
\examples{
#dataBRAVODayLabeller(x)
}
\keyword{BRAVO}
\keyword{CleanUp}
