% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{dataBRAVOSymptoms}
\alias{dataBRAVOSymptoms}
\title{BRAVO symptom extractor function}
\usage{
dataBRAVOSymptoms(x)
}
\arguments{
\item{x}{the BRAVO dataframe for extraction from SAPTotal as probably more accurate as patient entered (rather than from summary)}
}
\description{
This extracts the relevant symptoms from the impedance dataset
}
\examples{
#dataBRAVOSymptoms(x)
}
\keyword{Impedance}
\keyword{symptoms}
