% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{dataBRAVOTotal}
\alias{dataBRAVOTotal}
\title{dataBRAVOTotal
Extracts from the BRAVOTotal table
This acquires the first two BRAVO days data frmo the upper GI database}
\usage{
dataBRAVOTotal(channel)
}
\arguments{
\item{channel}{odbcConnectAccess connection defined at the start of this file from RODBC function (for windows)}
}
\description{
dataBRAVOTotal
Extracts from the BRAVOTotal table
This acquires the first two BRAVO days data frmo the upper GI database
}
\keyword{BRAVO}
\keyword{Extraction}
\keyword{total}
