% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{dataBT}
\alias{dataBT}
\title{dataBT
This cleans HRM data}
\usage{
dataBT(x, channel)
}
\arguments{
\item{x}{dataframe}
}
\description{
dataBT
This cleans HRM data
}
\examples{
dataBT(x)
}
\keyword{BT}
\keyword{CleanUp}
