% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagFunctions.R
\name{dataDiagClean}
\alias{dataDiagClean}
\title{Diag cleanup}
\usage{
dataDiagClean(x)
}
\arguments{
\item{x}{the Diag data frame}
}
\description{
Cleans up the Diag data
}
\examples{
#dataDiagClean(x)
}
\keyword{Diag}
\keyword{data}
