% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreathTests.R
\name{dataFructose}
\alias{dataFructose}
\title{dataFructose
This extracts the raw fructose data from the database}
\usage{
dataFructose(dataBT)
}
\arguments{
\item{dataBT}{the data frame}
}
\description{
dataFructose
This extracts the raw fructose data from the database
}
\examples{
#dataFructose(dataBT)
}
\keyword{Fructose}
