% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{dataHRM}
\alias{dataHRM}
\title{dataHRM
This acquires HRM data from the database}
\usage{
dataHRM(channel)
}
\arguments{
\item{channel}{odbcConnectAccess connection defined at the start of this file from RODBC function (for windows)}
}
\description{
dataHRM
This acquires HRM data from the database
}
\examples{
dataHRM(channel)
}
\keyword{HRM}
\keyword{acquisition}
\keyword{data}
