% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{dataImp2}
\alias{dataImp2}
\title{dataImp2
This acquires impedance data from the upper GI database main impedance table called Impedance2}
\usage{
dataImp2(channel)
}
\arguments{
\item{channel}{odbcConnectAccess connection defined at the start of this file from RODBC function (for windows)}
}
\description{
dataImp2
This acquires impedance data from the upper GI database main impedance table called Impedance2
}
\examples{
dataImp2(x)
}
\keyword{CleanUp}
\keyword{Imp}
