% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{dataImp_Symp}
\alias{dataImp_Symp}
\title{dataImp_Symp
This acquires impedance symptoms data from the upper GI database}
\usage{
dataImp_Symp(channel)
}
\arguments{
\item{channel}{odbcConnectAccess connection defined at the start of this file from RODBC function (for windows)}
}
\description{
dataImp_Symp
This acquires impedance symptoms data from the upper GI database
}
\examples{
#dataImp_Symp(channel)
}
\keyword{Extraction}
\keyword{Imp_Symp}
