% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{dataImpClean}
\alias{dataImpClean}
\title{dataImpClean
This extracts the symptoms}
\usage{
dataImpClean(x)
}
\arguments{
\item{x}{dataframe usually the standard impedance data}

\item{y}{the dataframe usually main Impedance table}
}
\description{
dataImpClean
This extracts the symptoms
}
\examples{
#dataImpClean(Impedance2)
}
\keyword{CleanUp}
\keyword{HRM}
