% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{dataImpSympClean}
\alias{dataImpSympClean}
\title{dataImpSympClean
This cleans the data from the ImpSymp table and the main Impedance table}
\usage{
dataImpSympClean(x)
}
\arguments{
\item{x}{dataframe usually the standard impedance data}
}
\description{
dataImpSympClean
This cleans the data from the ImpSymp table and the main Impedance table
}
\examples{
#dataImpSympClean(x)
}
\keyword{CleanUp}
\keyword{ImpSymp}
