% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{FunctionalHeartburn}
\alias{FunctionalHeartburn}
\title{A Cat Function}
\usage{
FunctionalHeartburn(x)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
