% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{GORD_AcidBRAVO}
\alias{GORD_AcidBRAVO}
\title{Diagnosis of GORD for BRAVO studies}
\usage{
GORD_AcidBRAVO(dd)
}
\arguments{
\item{x}{the impedance dataset for extraction}
}
\description{
This extracts whether the patient had a formal GORD diagnosis
This is based on the Acid exposure table but the rules are different for BRAVO results based on the day
}
\details{
The rules are that if a patient has a long acid exposure percentage (>4.2% total)
Or if there are a large number of reflux events (>73) - (not that the reflux episodes are acid only
Or if the the final report says pathological reflux or nocturnal (as the total may be normal) then the patient has a GORD diagnosis
}
\examples{
#GORD_AcidBRAVO(x)
}
\keyword{BRAVO}
\keyword{GORD}
\keyword{acid}
