% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{GORD_BravoWDAAndAverage}
\alias{GORD_BravoWDAAndAverage}
\title{Create Worst day and Average day analysis of BRAVO}
\usage{
GORD_BravoWDAAndAverage(x)
}
\arguments{
\item{x}{the impedance dataset for extraction}
}
\description{
This extracts whether the patient had a formal GORD diagnosis
This is based on the Acid exposure table. The rules are that if a patient has a long acid exposure percentage (>4.2% total)
Or if there are a large number of reflux events (>73) - field called
Or if the the final report says pathological reflux or nocturnal (as the total may be normal) then the patient has a GORD diagnosis
}
\examples{
#GORD_AcidImpImp(x)
}
\keyword{GORD}
\keyword{Impedance}
\keyword{acid}
