% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{HRMAndDiag}
\alias{HRMAndDiag}
\title{HRMAndDiag
This acquires HRM and associated diagnosis data from the upper GI database}
\usage{
HRMAndDiag(channel)
}
\arguments{
\item{channel}{odbcConnectAccess connection defined at the start of this file from RODBC function (for windows)}
}
\description{
HRMAndDiag
This acquires HRM and associated diagnosis data from the upper GI database
}
\examples{
HRMAndDiag(channel)
}
\keyword{Diag}
\keyword{HRM}
\keyword{and}
