% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{HRMAndSwallows}
\alias{HRMAndSwallows}
\title{HRMAndSwallows
This cleans HRM data}
\usage{
HRMAndSwallows(x, channel)
}
\arguments{
\item{x}{dataframe}
}
\description{
HRMAndSwallows
This cleans HRM data
}
\examples{
HRMAndSwallows(x)
}
\keyword{CleanUp}
\keyword{HRMAndSwallows}
