% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{HRMCleanUp}
\alias{HRMCleanUp}
\title{HRMCleanUp
This cleans HRM data}
\usage{
HRMCleanUp(x, channel)
}
\arguments{
\item{x}{dataframe}
}
\description{
HRMCleanUp
This cleans HRM data
}
\examples{
HRMCleanUp(x)
}
\keyword{CleanUp}
\keyword{HRM}
