% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotilityFunctions.R
\name{HRMCleanUp1}
\alias{HRMCleanUp1}
\title{HRMCleanUp1
This cleans HRM data by making sure all the data is in the correct format}
\usage{
HRMCleanUp1(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
HRMCleanUp1
This cleans HRM data by making sure all the data is in the correct format
}
\examples{
#HRMCleanUp1(x)
}
\keyword{CleanUp}
\keyword{HRM}
