% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{HypersensitiveOesophagus}
\alias{HypersensitiveOesophagus}
\title{hypersensitive oesophagus function}
\usage{
HypersensitiveOesophagus(x)
}
\arguments{
\item{x}{the impedance dataset for extraction}
}
\description{
Extracts all the patients with hypersensitive oesophagus
}
\examples{
#HypersensitiveOesophagus(x)
}
\keyword{hypersensitive}
\keyword{oesophagus}
