% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{MotilityTimeSeries}
\alias{MotilityTimeSeries}
\title{MotilityTimeSeries
Plots the Tests over time. Probably redundant}
\usage{
MotilityTimeSeries(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
MotilityTimeSeries
Plots the Tests over time. Probably redundant
}
\examples{
MotilityTimeSeries(x)

}
\keyword{HRM}
\keyword{Motility}
