% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{MyImpedanceDataWithHRM}
\alias{MyImpedanceDataWithHRM}
\title{MyImpedanceDataWithHRM
This merges HRM with Impedancedata}
\usage{
MyImpedanceDataWithHRM(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
MyImpedanceDataWithHRM
This merges HRM with Impedancedata
}
\examples{
MyImpedanceDataWithHRM(x)
}
\keyword{CleanUp}
\keyword{HRM}
