% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{SupragastricBelching}
\alias{SupragastricBelching}
\title{Supragastric belching function}
\usage{
SupragastricBelching(x)
}
\arguments{
\item{x}{the impedance dataset for extraction}
}
\description{
Extracts all the patients with supragastric belchng
}
\examples{
#SupragastricBelching(x)
}
\keyword{belching}
