% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotilityFunctions.R
\name{Symptom_Plot}
\alias{Symptom_Plot}
\title{Symptoms
This barcharts the symptoms}
\usage{
Symptom_Plot(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
Symptoms
This barcharts the symptoms
}
\examples{
#Symptoms(x)
}
\keyword{CleanUp}
\keyword{HRM}
