% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{SymptomExtractor}
\alias{SymptomExtractor}
\title{SymptomExtractor
This extracts the symptoms}
\usage{
SymptomExtractor(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
SymptomExtractor
This extracts the symptoms
}
\examples{
SymptomExtractor(x)

}
\keyword{CleanUp}
\keyword{HRM}
