% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{Symptoms}
\alias{Symptoms}
\title{Symptoms
This barcharts the symptoms}
\usage{
Symptoms(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
Symptoms
This barcharts the symptoms
}
\examples{
Symptoms(x)

}
\keyword{CleanUp}
\keyword{HRM}
