% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotilityFunctions.R
\name{SymptomsExtraction}
\alias{SymptomsExtraction}
\title{SymptomsExtraction}
\usage{
SymptomsExtraction(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{the column of interest}
}
\description{
SymptomsExtraction
This extracts the symptoms from a pre-specified column into its own column per symptom. Usually these are extracted from the
associated diagnosis table as the HRM main table doesnt keep free text stuff
}
\examples{
#SymptomsExtraction(x,IndicANDHx)
}
\keyword{Symptom}
