% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{SymptomsNoPlot}
\alias{SymptomsNoPlot}
\title{SymptomsNoPlot
This get the number symptoms. Very likely redundant}
\usage{
SymptomsNoPlot(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
SymptomsNoPlot
This get the number symptoms. Very likely redundant
}
\examples{
SymptomsNoPlot(x)

}
\keyword{CleanUp}
\keyword{HRM}
