% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemographicFunctions.R
\name{testMerge}
\alias{testMerge}
\title{testMerge
This extracts the symptoms}
\usage{
testMerge(test1, test2, test1VisitDate, test2VisitDate)
}
\arguments{
\item{test1}{dataframe}

\item{test2}{dataframe}

\item{test1VisitDate}{VisitDate from the first test}

\item{test2VisitDate}{VisitDate from the second test}
}
\description{
testMerge
This extracts the symptoms
}
\examples{
#testMerge(test1,test2,test1VisitDate,test2VisitDate)
}
\keyword{CleanUp}
\keyword{HRM}
