% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{AcidSubtypes}
\alias{AcidSubtypes}
\title{Acid subtype extractor
This creates the composite score and then subcategorises the reflux ie if acid reflux then it is recumbent vs upright vs postprandial etc. (postprandial to be done)}
\usage{
AcidSubtypes(x)
}
\arguments{
\item{x}{the impedance dataset for extraction}
}
\description{
Acid subtype extractor
This creates the composite score and then subcategorises the reflux ie if acid reflux then it is recumbent vs upright vs postprandial etc. (postprandial to be done)
}
\examples{
#AcidSubtypes(x)
}
\keyword{Impedance}
\keyword{acid}
