% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotilityFunctions.R
\name{BasicBoxplots}
\alias{BasicBoxplots}
\title{BasicBoxplots
This boxplots HRM measurements. Very likely redundant}
\usage{
BasicBoxplots(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
BasicBoxplots
This boxplots HRM measurements. Very likely redundant
}
\examples{
#BasicBoxplots(x)
}
\keyword{CleanUp}
\keyword{HRM}
