% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{BRAVO}
\alias{BRAVO}
\title{BRAVO
This acquires BRAVO data from the upper GI database}
\usage{
BRAVO(channel)
}
\arguments{
\item{channel}{odbcConnectAccess connection defined at the start of this file from RODBC function (for windows)}
}
\description{
BRAVO
This acquires BRAVO data from the upper GI database
}
\examples{
BRAVO(channel)
}
\keyword{BRAVO}
\keyword{CleanUp}
