% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreathTests.R
\name{dataBT}
\alias{dataBT}
\title{dataBT
This extracts the breath test data}
\usage{
dataBT(channel)
}
\arguments{
\item{channel}{the data base query}
}
\description{
dataBT
This extracts the breath test data
}
\examples{
#dataBT(channel)
}
\keyword{Fructose}
