% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{dataBT}
\alias{dataBT}
\title{dataBT
This acquires Breath test data from the upper GI database}
\usage{
dataBT(channel)
}
\arguments{
\item{channel}{odbcConnectAccess connection defined at the start of this file from RODBC function (for windows)}
}
\description{
dataBT
This acquires Breath test data from the upper GI database
}
\examples{
dataBT(channel)
}
\keyword{BT}
\keyword{CleanUp}
