% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{dataFructose}
\alias{dataFructose}
\title{dataFructose
This extracts the symptoms}
\usage{
dataFructose(dataBT)
}
\arguments{
\item{x}{the data frame}
}
\description{
dataFructose
This extracts the symptoms
}
\examples{
dataImpClean(dataBT)
}
\keyword{Fructose}
