% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{dataHRM}
\alias{dataHRM}
\title{dataHRM
This cleans HRM data}
\usage{
dataHRM(x, channel)
}
\arguments{
\item{x}{dataframe}
}
\description{
dataHRM
This cleans HRM data
}
\examples{
dataHRM(x)
}
\keyword{CleanUp}
\keyword{HRM}
