% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{dataImp2}
\alias{dataImp2}
\title{dataImp2
This cleans HRM data}
\usage{
dataImp2(x, channel)
}
\arguments{
\item{x}{dataframe}
}
\description{
dataImp2
This cleans HRM data
}
\examples{
dataImp2(x)
}
\keyword{CleanUp}
\keyword{Imp}
