% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{dataImp_Symp}
\alias{dataImp_Symp}
\title{dataImp_Symp
This cleans HRM data}
\usage{
dataImp_Symp(x, channel)
}
\arguments{
\item{x}{dataframe}
}
\description{
dataImp_Symp
This cleans HRM data
}
\examples{
dataImp_Symp(x)
}
\keyword{CleanUp}
\keyword{Imp_Symp}
