% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{dataImpClean}
\alias{dataImpClean}
\title{dataImpClean
This extracts the symptoms}
\usage{
dataImpClean(x, y)
}
\arguments{
\item{x}{dataframe usually the standard impedance data}

\item{y}{the dataframe usually the symptom data}
}
\description{
dataImpClean
This extracts the symptoms
}
\examples{
#dataImpClean(dataImp2,dataImp_Symp)
}
\keyword{CleanUp}
\keyword{HRM}
