% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R, R/Physiology.R
\name{dataImpClean}
\alias{dataImpClean}
\alias{dataImpClean}
\title{A Cat Function}
\usage{
dataImpClean(x, y)

dataImpClean(x, y)
}
\arguments{
\item{x}{dataframe usually the standard impedance data}

\item{y}{the dataframe usually the symptom data}

\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.

dataImpClean
This extracts the symptoms
}
\examples{
cat_function()
dataImpClean(dataImp2,dataImp_Symp)
}
\keyword{CleanUp}
\keyword{HRM}
\keyword{cats}
