% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{dataImpSymptoms}
\alias{dataImpSymptoms}
\title{A Cat Function}
\usage{
dataImpSymptoms(x)
}
\arguments{
\item{x}{the impedance dataframe for extraction}
}
\description{
This extracts the relevant symptoms from the impedance dataset
}
\examples{
#dataImpSymptoms(x)
}
\keyword{Impedance}
\keyword{symptoms}
