% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{dataImpWhole}
\alias{dataImpWhole}
\title{dataImpWhole
This cleans HRM data}
\usage{
dataImpWhole(x, channel)
}
\arguments{
\item{x}{dataframe}
}
\description{
dataImpWhole
This cleans HRM data
}
\examples{
dataImpWhole(x)
}
\keyword{CleanUp}
\keyword{ImpWhole}
