% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{dataImpWhole}
\alias{dataImpWhole}
\title{dataImpWhole
This merges the main impedance and the impedance symptoms data frames together
It relies on the acquisition of the data from the functions dataImp2 and dataImp_Symp}
\usage{
dataImpWhole(dataImp2, dataImp_Symp)
}
\arguments{
\item{channel}{odbcConnectAccess connection defined at the start of this file from RODBC function (for windows)}
}
\description{
dataImpWhole
This merges the main impedance and the impedance symptoms data frames together
It relies on the acquisition of the data from the functions dataImp2 and dataImp_Symp
}
\examples{
dataImp2<-dataImp2(channel),
dataImp_Symp<-dataImp_Symp(channel)
dataImpWhole(dataImp2,dataImp_Symp)
}
\keyword{CleanUp}
\keyword{ImpWhole}
