% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{FunctionalHeartburn}
\alias{FunctionalHeartburn}
\title{FunctionalHeartburn function}
\usage{
FunctionalHeartburn(x)
}
\arguments{
\item{x}{the impedance dataset for extraction}
}
\description{
Extracts all the patients with Functional Heartburn
}
\examples{
#FunctionalHeartburn(x)
}
\keyword{Functional}
\keyword{Heartburn}
