% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{HRMAndDiag}
\alias{HRMAndDiag}
\title{HRMAndDiag
This cleans HRM data}
\usage{
HRMAndDiag(x, channel)
}
\arguments{
\item{x}{dataframe}
}
\description{
HRMAndDiag
This cleans HRM data
}
\examples{
HRMAndDiag(x)
}
\keyword{Diag}
\keyword{HRM}
\keyword{and}
