% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{HRMAndSwallows}
\alias{HRMAndSwallows}
\title{HRMAndSwallows
This acquires HRMSwallow data from the upper GI database}
\usage{
HRMAndSwallows(channel)
}
\arguments{
\item{channel}{odbcConnectAccess connection defined at the start of this file from RODBC function (for windows)}
}
\description{
HRMAndSwallows
This acquires HRMSwallow data from the upper GI database
}
\examples{
HRMAndSwallows(channel)
}
\keyword{HRMAndSwallows}
\keyword{acquisition}
