% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotilityFunctions.R, R/Physiology.R
\name{HRMCleanUp1}
\alias{HRMCleanUp1}
\alias{HRMCleanUp1}
\title{A Cat Function}
\usage{
HRMCleanUp1(x)

HRMCleanUp1(x)
}
\arguments{
\item{x}{dataframe}

\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.

HRMCleanUp1
This cleans HRM data by making sure all the data is in the correct format
}
\examples{
cat_function()
HRMCleanUp1(x)
}
\keyword{CleanUp}
\keyword{HRM}
\keyword{cats}
