% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physiology.R
\name{MyImpedanceDataWithHRM}
\alias{MyImpedanceDataWithHRM}
\title{MyImpedanceDataWithHRM
This merges HRM with Impedancedata. This relies on the two previous functions to acquire the data
first from the upper GI database, namely dataImpWhole and dataHRM}
\usage{
MyImpedanceDataWithHRM(dataImpWhole, dataHRM)
}
\arguments{
\item{dataImpWhole}{dataframe}

\item{dataHRM}{dataframe}
}
\description{
MyImpedanceDataWithHRM
This merges HRM with Impedancedata. This relies on the two previous functions to acquire the data
first from the upper GI database, namely dataImpWhole and dataHRM
}
\examples{
dataImpWhole<-dataImpWhole(channel)
dataHRM<-dataHRM(channel)
MyImpedanceDataWithHRM(dataImpWhole,data)
}
\keyword{CleanUp}
\keyword{HRM}
