% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotilityFunctions.R
\name{MyImpedanceDataWithHRM}
\alias{MyImpedanceDataWithHRM}
\title{Merge HRM and Impedance
This merges the overall impedance dataset with the HRM dataset}
\usage{
MyImpedanceDataWithHRM(x, y)
}
\arguments{
\item{x}{the whole impedance dataset}

\item{y}{the whole HRM dataset}
}
\description{
Merge HRM and Impedance
This merges the overall impedance dataset with the HRM dataset
}
\examples{
#dataHRM<-dataHRM(channel)
#dataImp2<-dataImp2(channel)
#dataImp_Symp<-dataImp_Symp(channel)

#dataImpWhole<-dataImpWhole(dataImp2,dataImp_Symp)
#MyImpedanceDataWithHRM(dataImpWhole,dataHRM)
}
\keyword{HRMAndSwallows}
\keyword{acquisition}
