% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpedanceFunctions.R
\name{SupragastricBelching}
\alias{SupragastricBelching}
\title{Supragastric belching function}
\usage{
SupragastricBelching(x)
}
\arguments{
\item{x}{the impedance merged with whole report dataset for extraction}
}
\description{
Extracts all the patients with supragastric belching- can only be done once the whole report
#has been merged with the impedance results
}
\examples{
#SupragastricBelching(x)
}
\keyword{belching}
